requirements.txt
asgiref==3.8.1
attrs==23.2.0
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
drf-spectacular==0.27.2
drf-yasg==1.21.7
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
packaging==24.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.0
rpds-py==0.18.0
setuptools==69.5.1
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
